+++ wallabag.Api.Tests/Credentials.cs
@@ -6,7 +6,7 @@
 
         public const string wallabagUrl = "https://wallabag.jlnostr.de";
         public const string username = "wallabag";
-        public const string password = "wallabag";
+        public const string password = "g7idJ8jzpaUQfakz19mqkiDbwrot55f6";
         public const string clientId = "4_1bblj107c7j4g8ss44co488sk80k8koscwowwo0wcsg08wk8w8";
         public const string clientSecret = "rzurdc5w7u8og408kwcs4wc4kwgoo808ggocwsgg0gs80480c";
     }