+++ wallabag.Api/Components/Modify.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using wallabag.Api.Models;
 
@@ -12,13 +13,13 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="itemId">The item id.</param>
         /// <returns>True, if the action was successful.</returns>
-        public async Task<bool> ArchiveAsync(int itemId)
+        public async Task<bool> ArchiveAsync(int itemId, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (itemId == 0)
                 throw new ArgumentNullException(nameof(itemId));
 
-            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", new Dictionary<string, object>() { ["archive"] = true.ToInt() });
-            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString);
+            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", cancellationToken, new Dictionary<string, object>() { ["archive"] = true.ToInt() });
+            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString, cancellationToken);
 
             return item?.IsRead == true;
         }
@@ -28,14 +29,14 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="itemId">The item id.</param>
         /// <returns>True, if the action was successful.</returns>
-        public async Task<bool> UnarchiveAsync(int itemId)
+        public async Task<bool> UnarchiveAsync(int itemId, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (itemId == 0)
                 throw new ArgumentNullException(nameof(itemId));
 
-            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", new Dictionary<string, object>() { ["archive"] = false.ToInt() });
-            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString);
-            
+            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", cancellationToken, new Dictionary<string, object>() { ["archive"] = false.ToInt() });
+            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString, cancellationToken);
+
             return item?.IsRead == false;
         }
 
@@ -44,13 +45,13 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="itemId">The item id.</param>
         /// <returns>True, if the action was successful.</returns>
-        public async Task<bool> FavoriteAsync(int itemId)
+        public async Task<bool> FavoriteAsync(int itemId, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (itemId == 0)
                 throw new ArgumentNullException(nameof(itemId));
 
-            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", new Dictionary<string, object>() { ["starred"] = true.ToInt() });
-            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString);
+            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", cancellationToken, new Dictionary<string, object>() { ["starred"] = true.ToInt() });
+            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString, cancellationToken);
 
             return item?.IsStarred == true;
         }
@@ -60,13 +61,13 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="itemId">The item id.</param>
         /// <returns>True, if the action was successful.</returns>
-        public async Task<bool> UnfavoriteAsync(int itemId)
+        public async Task<bool> UnfavoriteAsync(int itemId, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (itemId == 0)
                 throw new ArgumentNullException(nameof(itemId));
 
-            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", new Dictionary<string, object>() { ["starred"] = false.ToInt() });
-            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString);
+            var jsonString = await ExecuteHttpRequestAsync(HttpRequestMethod.Patch, $"/entries/{itemId}", cancellationToken, new Dictionary<string, object>() { ["starred"] = false.ToInt() });
+            var item = await ParseJsonFromStringAsync<WallabagItem>(jsonString, cancellationToken);
 
             return item?.IsStarred == false;
         }
@@ -76,12 +77,12 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="itemId">The item id.</param>
         /// <returns>True, if the action was successful.</returns>
-        public async Task<bool> DeleteAsync(int itemId)
+        public async Task<bool> DeleteAsync(int itemId, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (itemId == 0)
                 throw new ArgumentNullException(nameof(itemId));
 
-            await ExecuteHttpRequestAsync(HttpRequestMethod.Delete, $"/entries/{itemId}");
+            await ExecuteHttpRequestAsync(HttpRequestMethod.Delete, $"/entries/{itemId}", cancellationToken);
 
             //TODO: Better check for actual result.
             return true;
@@ -92,34 +93,34 @@ namespace wallabag.Api
         /// </summary>
         /// <param name="item">The item that should be marked as read.</param>
         /// <returns>True, if the action was successful.</returns>
-        public Task<bool> ArchiveAsync(WallabagItem item) => ArchiveAsync(item.Id);
+        public Task<bool> ArchiveAsync(WallabagItem item, CancellationToken cancellationToken = default(CancellationToken)) => ArchiveAsync(item.Id, cancellationToken);
 
         /// <summary>
         /// Unmarks an item as read.
         /// </summary>
         /// <param name="item">The item that should be unmarked as read.</param>
         /// <returns>True, if the action was successful.</returns>
-        public Task<bool> UnarchiveAsync(WallabagItem item) => UnarchiveAsync(item.Id);
+        public Task<bool> UnarchiveAsync(WallabagItem item, CancellationToken cancellationToken = default(CancellationToken)) => UnarchiveAsync(item.Id, cancellationToken);
 
         /// <summary>
         /// Marks an item as starred.
         /// </summary>
         /// <param name="item">The item that should be marked as starred.</param>
         /// <returns>True, if the action was successful.</returns>
-        public Task<bool> FavoriteAsync(WallabagItem item) => FavoriteAsync(item.Id);
+        public Task<bool> FavoriteAsync(WallabagItem item, CancellationToken cancellationToken = default(CancellationToken)) => FavoriteAsync(item.Id, cancellationToken);
 
         /// <summary>
         /// Unmarks an item as starred.
         /// </summary>
         /// <param name="item">The item that should be unmarked as starred.</param>
         /// <returns>True, if the action was successful.</returns>
-        public Task<bool> UnfavoriteAsync(WallabagItem item) => UnfavoriteAsync(item.Id);
+        public Task<bool> UnfavoriteAsync(WallabagItem item, CancellationToken cancellationToken = default(CancellationToken)) => UnfavoriteAsync(item.Id, cancellationToken);
 
         /// <summary>
         /// Deletes an item PERMANENTLY.
         /// </summary>
         /// <param name="item">The item that should be deleted.</param>
         /// <returns>True, if the action was successful.</returns>
-        public Task<bool> DeleteAsync(WallabagItem item) => DeleteAsync(item.Id);
+        public Task<bool> DeleteAsync(WallabagItem item, CancellationToken cancellationToken = default(CancellationToken)) => DeleteAsync(item.Id, cancellationToken);
     }
 }